registrasi/exceptionhandler.py
from rest_framework.views import exception_handler

def custom_exception_handler(exc, context):
    # Call REST framework's default exception handler first,
    # to get the standard error response.
    response = exception_handler(exc, context)

    # Now add the HTTP status code to the response.
    if response.status_code==403 is not None:
        response.data['detail'] = 'Halo hekelmen'

    return response