requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.9.4
gunicorn==19.9.0
httpie==1.0.2
idna==2.8
psycopg2==2.8.2
Pygments==2.4.2
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
